fix(platform_api): fix int64 type assertion panic in GetMergeability
In todosCompletionBlock, the uid=0(root) gid=0(root) groups=0(root) key is stored as int64 (t.ID),
but GetMergeability asserted it as float64, causing a panic
and HTTP 500 on /issues/current/mergeability.

+++ apps/hangrix/internal/modules/platform_api/service/api.go
@@ -739,7 +739,7 @@ func (s *APIService) GetMergeability(ctx context.Context, p *apidomain.Actor) (a
 			hint = block
 			for _, t := range incTodos {
 				incompleteTodos = append(incompleteTodos, apiTodo{
-					ID:      int64(t["id"].(float64)),
+					ID:      t["id"].(int64),
 					Content: t["content"].(string),
 					Status:  t["status"].(string),
 				})
